face-detection/requirements.txt
uvicorn
fastapi
requests
pillow
python-dotenv
mtcnn
tensorflow
opencv-python
